finished todo list

+++ src/main/kotlin/dev/jombi/knuserver/entity/Review.kt
@@ -9,7 +9,7 @@ data class Review(
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "placeId")
     val place: Place,
-    val disabilityFacility: Byte,
+    val disabilityFacility: Int,
     @Column(length = 500)
     val comment: String,
 )
